Made EventIngestionClient, A subclass of discord.Client
Subclassed the discord.Client to make a new class that replaces the need of the make_client function. In doing this it removed all the @client.async_event decorators from the client functions. It also allowed the run function to be replaced so it could include the token and if it was a bot/selfbot.